Restructuring the WidgetForm component and creating the FeedbackTypeStep component
